requirements.txt
Pillow
tqdm
numpy
openai
python-dotenv